DRY: persistentLink implementation
